requirements.txt
blinker==1.9.0
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.2
flask-cors==6.0.2
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
pycparser==2.23
pypdfium2==5.4.0
python-docx==1.2.0
typing_extensions==4.15.0
Werkzeug==3.1.5
zipp==3.23.0
